Move `snippets` and related code to `applications`.
This functionality is only used in `applications` for defining some reused text
in some of the forms in that app. We have no plans to use it anywhere else. It
may as well live in that app, rather than cluttering up the top level directory
and `jobserver`.
