+++ applications/snippets/__init__.py
@@ -1,0 +1,1 @@
+from .snippets import render_snippet as render_snippet
