+++ applications/snippets/content/1-rubric.md
