+++ applications/snippets/content/3-rubric.md
