+++ applications/snippets/content/6-rubric.md
