+++ applications/snippets/content/application.md
