+++ applications/snippets/content/description-help_text.md
