+++ applications/snippets/content/previous_experience_with_ehr_help_text.md
