+++ applications/snippets/content/references-footer.md
