+++ applications/snippets/content/researcher-details.md
