+++ applications/snippets/content/short-data-report-rubric.md
