+++ applications/snippets/content/terms.md
