+++ applications/snippets/snippets.py
@@ -4,7 +4,7 @@ from markdown import markdown
 
 
 def render_snippet(key):
-    path = settings.BASE_DIR / "snippets" / f"{key}.md"
+    path = settings.BASE_DIR / "applications" / "snippets" / "content" / f"{key}.md"
     text = path.read_text()
     content = markdown(text)
     return mark_safe(content)
