+++ tests/unit/applications/test_snippets.py
@@ -2,10 +2,11 @@ from pathlib import Path
 
 from django.test import override_settings
 
-from jobserver.templatetags.snippet import snippet
+from applications.snippets import render_snippet
 
 
 @override_settings(BASE_DIR=Path(__file__).parents[1])
-def test_snippet():
-    output = snippet("test")
+def test_render_snippet():
+    output = render_snippet("test")
+    assert hasattr(output, "__html__")
     assert output == "<h1>A Title</h1>\n<p>Some content</p>"
